working on LongBench evaluation

+++ kv_quant/run_ppl.py
@@ -57,7 +57,7 @@ def eval_ppl(model, tokenizer, args, device="cuda"):
 
             results["wikitext"] = ppl.item()
         elif task_eval == "c4":
-            model_net = args.model_name_or_path.split('/')[-1]
+            model_net = model_name_or_path.split('/')[-1]
             model_family = '_'.join(model_net.lower().split('-')[:-1])
             model.seq_len = args.seq_len
 
@@ -114,18 +114,21 @@ if __name__ == '__main__':
     parser.add_argument('--datasets', type=lambda s: [item for item in s.split(',')], default=['wikitext'], help="Task to be evaled")
     parser.add_argument('--seq_len', type=int, help='sequence length for ppl evaluation', default=2048)
     parser.add_argument("--verbose", action="store_true", help="Whether to print verbose information or not.")
-    parser.add_argument("--output_dir", type=str, default="results/lm_eval", help="output directory")
+    parser.add_argument("--output_dir", type=str, default="results/ppl", help="output directory")
     args = parser.parse_args()  
+    quant_config = get_quant_config(args)
+    model_name = args.model_name
+
+    model2path = json.load(open("config/model2path.json", "r"))
+    model_name_or_path = model2path[model_name]
 
     logger.remove()
     logger.add(lambda msg: tqdm.write(msg, end=""), colorize=True, level="INFO" if not args.verbose else "DEBUG")
-    
-    quant_config = get_quant_config(args)
-    logger.info(f"Start evaluating ppl...")
-    logger.info(f"*model: {args.model_name_or_path}")
-    logger.info(f"*datasets: {args.datasets}")
-    logger.info(f"*sequence length {args.seq_len}")
-    logger.info(f"Start evaluating with the following configurations:")
+    logger.info(f"#################### Model Info ####################")
+    logger.info(f"* Model: {model_name_or_path}")
+    logger.info(f"* Datasets: {args.datasets}")
+    logger.info(f"* sequence length {args.seq_len}")
+    logger.info(f"#################### Start evaluating ppl with the following configurations: ####################")
     logger.info(f"* Bench compression!!!")
     logger.info(f"* KV-cache quantization method: {args.kv_quant_method}")
     logger.info(f"* Key bits: {args.k_bits}")
@@ -137,21 +140,27 @@ if __name__ == '__main__':
     logger.info(f"* Apply key bias?: {args.apply_k_bias}")
     logger.info(f"* Apply key scale?: {args.apply_k_scale}")
 
-    # Create directory if it doesn't exist
-    output_dir = args.output_dir
+    logger.info("#################### Creating output directory ... ####################")
+    output_dir = os.path.join(args.output_dir, model_name)
     os.makedirs(output_dir, exist_ok=True)
-    output_file_name = f"{args.kv_quant_method}-kbits_{args.k_bits}-vbits_{args.v_bits}-kgs_{args.k_group_size}-vgs_{args.v_group_size}-res_{args.kv_residual_len}-bias_{args.apply_k_bias}-scale_{args.apply_k_scale}"
+    output_file_name = f"{args.kv_quant_method}-kbits_{args.k_bits}-vbits_{args.v_bits}-kgs_{args.k_group_size}-vgs_{args.v_group_size}-res_{args.kv_residual_len}-pbitspf_{args.p_bits_pf}-pbitsdc_{args.p_bits_dc}-bias_{args.apply_k_bias}-scale_{args.apply_k_scale}"
     output_file_path = os.path.join(output_dir, f"{output_file_name}.txt")
+    # check if result file exists
+    print(output_file_path)
+    if os.path.isfile(output_file_path):
+        print(f'Found existing output file {output_file_name} for this experiment. Exit!')
+        exit()
     
-    logger.info("Loading model and tokenizer...")
-    model, tokenizer = load_model_and_tokenizer(args.model_name_or_path, quant_config=quant_config)
-    logger.info("Start running perplexity evaluation...")
+    logger.info("#################### Loading model and tokenizer ... ####################")
+    model, tokenizer = load_model_and_tokenizer(model_name_or_path, quant_config=quant_config, use_fp16=args.use_fp16)
+
+    logger.info("#################### Start running perplexity evaluation ... ####################")
     res = eval_ppl(model, tokenizer, args)
 
     # Save results to JSON file
     with open(output_file_path, "w") as f:
         for dataset, ppl in res.items():
             logger.info(f"{dataset} PPL: {ppl}")
-            f.write(f"{dataset} PPL: {ppl}\n")
+            f.write(f"{dataset.ljust(10)} PPL: {ppl}\n")
     
     print(f"Results saved to {output_file_path}")
